+++ esphome/components/teleinfo/teleinfo.cpp
@@ -116,7 +116,7 @@ void TeleInfo::loop() {
        * 0xa | Tag | 0x9 | Timestamp | 0x9 | 0x9 | CRC | 0xd
        *
        */
-      while ((buf_finger = static_cast<char *>(memchr(buf_finger, (int) 0xa, buf_index_ - 1))) &&
+      while ((buf_finger = static_cast<char *>(memchr(buf_finger, 0xa, buf_index_ - 1))) &&
              ((buf_finger - buf_) < buf_index_)) {  // NOLINT(clang-diagnostic-sign-compare)
         /*
          * Make sure timesamp is nullified between each tag as some tags don't
